+++ README.md
@@ -26,7 +26,6 @@
 1. git 레포지토리를 fork하여 작성한 후 PR을 보내세요
 2. 리팩토링 의도와 구조에 대한 간략한 설명을 PR본문에 적어주세요
 
-
 ## 기타
 
 - 디자인은 마음대로 변경하셔도 됩니다.
